ai/requirements.txt
langchain-groq
python-dotenv
langchain-google-community\[gmail\]
langgraph